moved some horizon functions to tools and fixed naming. Also wrote 2 tests for new irradiance functions.
